pyproject.toml
[tool.poetry]
name = "minerva-backend"
version = "0.0.1"
description = "Backend team repository of Minerva Learning Management System (LMS). Developed by a group of students of Universidad Nacional de Colombia using Django."
authors = ["Carolina Alvarez Murillo <caalvarezmu@unal.edu.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.20.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = true
in-project = true